backtesting/requirements.txt
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
db-dtypes==1.4.3
fonttools==4.58.4
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-bigquery==3.34.0
google-cloud-core==2.4.3
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.73.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.3.1
orjson==3.10.18
packaging==25.0
pandas==2.3.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
rsa==4.9.1
seaborn==0.13.2
six==1.17.0
sqlitedict==2.1.0
toml==0.10.2
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
zelos-demeter==1.0.1
python-dotenv==1.1.0
